backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
flask-socketio==5.3.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
resend==2.15.0
cryptography==41.0.7
Pillow==10.1.0
PyPDF2==3.0.1
qrcode==7.4.2
python-dateutil==2.8.2
gunicorn==21.2.0
eventlet==0.33.3
